django/byutest/byutest/serializers.py
from rest_framework import serializers
from django.contrib.auth.models import User
from django.contrib.auth.hashers import check_password
from django.contrib.auth.password_validation import validate_password
from django.core.exceptions import ValidationError as DjangoValidationError
from byutest.models import *
from byutest.views.helpers import check_inappropriate_language
import logging
from django.db import transaction
import logging

logger = logging.getLogger(__name__)

# ----------------------------------------------------------------------

logging.basicConfig(level=logging.INFO)

# ----------------------------------------------------------------------

