django/requirements.txt
boto3
Django
django-environ
djangorestframework
djangorestframework-simplejwt
django-cors-headers~=3.13.0
drf-jwt~=1.19.2
dj-rest-auth
django-allauth
mysqlclient
pandas
uwsgi